Parse all possible types of CellInfo

+++ app/src/main/java/com/cellwatch/data/network/model/NetworkCell.kt
@@ -12,7 +12,7 @@ data class NetworkCell(
     val timestamp: Instant? = null,
 
     @SerialName("cell_id")
-    val cellId: Int? = null,
+    val cellId: Long? = null,
 
     @SerialName("physical_cell_id")
     val physicalCellId: Int? = null,
